Extract ai provider label getter to own lib module
