+++ assets/js/common/AIProviderLabel/AIProviderLabel.jsx
@@ -1,32 +1,11 @@
 import React from 'react';
-import { get } from 'lodash';
 
 import { EOS_HELP } from 'eos-icons-react';
 
-import GeminiIcon from '@static/gemini-logo.svg';
-import OpenAIIcon from '@static/openai-logo.svg';
-import AnthropicIcon from '@static/anthropic-logo.svg';
-
-const providerRenderingConfig = {
-  googleai: {
-    icon: GeminiIcon,
-    label: 'Google Gemini',
-  },
-  openai: {
-    icon: OpenAIIcon,
-    label: 'OpenAI GPT',
-  },
-  anthropic: {
-    icon: AnthropicIcon,
-    label: 'Anthropic Claude',
-  },
-};
-
-const getProviderLabel = (provider) =>
-  get(providerRenderingConfig, [provider, 'label'], provider);
+import { getProviderLabel, getProviderIcon } from '@lib/ai';
 
 function ProviderIcon({ provider }) {
-  const providerIcon = get(providerRenderingConfig, [provider, 'icon']);
+  const providerIcon = getProviderIcon(provider);
 
   if (providerIcon) {
     return (
